+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- multipart file upload support
+
 ### Changes
 
 - update mex-common to 0.33.0